+++ flake.nix
@@ -6,11 +6,16 @@
   inputs."overlays/paperwm".url = path:overlays/paperwm;
   inputs."overlays/paperwm".inputs.flake-utils.follows = "flake-utils-plus/flake-utils";
   inputs."overlays/paperwm".inputs.nixpkgs.follows = "nixpkgs";
+  inputs."overlays/nix".url = path:overlays/nix;
+  inputs."overlays/nix".inputs.inheritedNix.follows = "nix";
+  inputs."overlays/nix".inputs.flake-utils.follows = "flake-utils-plus/flake-utils";
+  inputs."overlays/nix".inputs.nixpkgs.follows = "nixpkgs";
+  inputs.nix.inputs.nixpkgs.follows = "nixpkgs";
 
-  outputs = inputs@{ self, nixpkgs, home-manager, flake-utils-plus, ... }:
+  outputs = inputs@{ self, nix, nixpkgs, home-manager, flake-utils-plus, ... }:
 
   let
-    inherit (builtins) hasAttr filter getAttr readDir;
+    inherit (builtins) attrNames hasAttr filter getAttr readDir;
     inherit (nixpkgs.lib) attrValues foldr filterAttrs getAttrFromPath hasSuffix mapAttrs' mapAttrsToList mkIf nameValuePair recursiveUpdate removeSuffix;
 
     importFrom = path: filename: import (path + ("/" + filename));
@@ -19,10 +24,28 @@
     regularOverlays = filterAttrs (name: _: hasSuffix ".nix" name) (readDir ./overlays);
     importedRegularOverlays = mapAttrsToList importOverlay regularOverlays;
 
-    flakeOverlays = builtins.attrNames (filterAttrs (_: type: type == "directory") (readDir ./overlays));
+    flakeOverlays = attrNames (filterAttrs (_: type: type == "directory") (readDir ./overlays));
     importedFlakeOverlays = map (name: getAttrFromPath [ "overlays/${name}" "overlay" ] inputs) flakeOverlays;
 
-    pkgs = import nixpkgs {
+    patches = let
+      inherit (import nixpkgs { system = "x86_64-linux"; }) fetchpatch;
+    in [
+      (fetchpatch {
+        name = "samba-open-firewall-ports-by-default.patch";
+        url = "https://github.com/NixOS/nixpkgs/commit/4f704e88b8c37d4514dc0a669986bc47ac2e86f0.patch";
+        sha256 = "sha256-EipMNHcqZitoGURYogz5u1AESSWld/ET4DdGzuVNWqE=";
+      })
+      # Fixes building Steam
+      (fetchpatch {
+        name = "tracker-patch-failing-test-due-to-float-comparison.patch";
+        url = "https://github.com/NixOS/nixpkgs/commit/b0ed9c8372686a513c947c204956e7a28e7e7147.patch";
+        sha256 = "sha256-j/cs4wK1twKSGHFoWtZI1g1IdWqVO0Dt6vSaADB64eo=";
+      })
+    ];
+
+    patchedNixpkgs = flake-utils-plus.lib.patchChannel "x86_64-linux" nixpkgs patches;
+
+    pkgs = import patchedNixpkgs {
       system = "x86_64-linux";
       config.allowUnfree = true;
       overlays = importedRegularOverlays ++ importedFlakeOverlays;
@@ -43,8 +66,8 @@
       home = [ ./home.nix ./hosts/${host}/home.nix ] ++ hmModules;
 
       configRevision = {
-        full = if (self ? rev) then self.rev else self.dirtyRev;
-        short = if (self ? rev) then self.shortRev else self.dirtyShortRev;
+        full = if (self ? rev) then self.rev else if (self ? dirtyRev) then self.dirtyRev else "dirty-inputs";
+        short = if (self ? rev) then self.shortRev else if (self ? dirtyRev) then self.dirtyShortRev else "dirty-inputs";
       };
     in {
       # nix build ~/.config/nixpkgs#nixosConfigurations.phi-nixos.config.system.build.toplevel